feat: cleaned code + improved performance
