+++ src/plugins/built-in/globalSearch/docs/client-vector-search.md
