+++ src/plugins/built-in/globalSearch/src/components/Calculator.svelte
