+++ src/plugins/built-in/globalSearch/src/core/commands.ts
