+++ src/plugins/built-in/globalSearch/src/core/styles.css
