+++ src/plugins/built-in/globalSearch/src/core/types.ts
