+++ src/plugins/built-in/globalSearch/src/indexing/actions.ts
