+++ src/plugins/built-in/globalSearch/src/indexing/db.ts
