+++ src/plugins/built-in/globalSearch/src/indexing/indexer.ts
