+++ src/plugins/built-in/globalSearch/src/indexing/jobs.ts
