+++ src/plugins/built-in/globalSearch/src/indexing/renderComponents.ts
