+++ src/plugins/built-in/globalSearch/src/indexing/types.ts
