+++ src/plugins/built-in/globalSearch/src/indexing/worker/vectorWorkerManager.ts
