+++ src/plugins/built-in/globalSearch/src/search/vector/vectorTypes.ts
