+++ src/plugins/built-in/globalSearch/src/utils/dynamicItems.ts
