+++ src/plugins/built-in/globalSearch/src/utils/highlight.ts
